simplifies command, removes bdv dependency now useless, split Command in 2 by first parsing the file then launching the interactive command
