+++ docs/ch/epfl/biop/operetta/commands/OperettaImporter.html
@@ -110,7 +110,7 @@ var activeTableTab = "activeTableTab";
 <li>org.scijava.command.InteractiveCommand</li>
 <li>
 <ul class="inheritance">
-<li>ch.epfl.biop.operetta.commands.OperettaImporter</li>
+<li>ch.epfl.biop.operetta.commands.OperettaImporterInteractive</li>
 </ul>
 </li>
 </ul>
